+++ src/dash-panels/_allPanels.ts
@@ -19,7 +19,7 @@ import text from './text.vue'
 import transit from './transit.vue'
 import vega from './vega.vue'
 import video from './video.vue'
-import overview from './overview.vue'
+import tile from './tile.vue'
 
 // full-screen map visualizations:
 import carriers from './carriers.vue'
@@ -34,6 +34,7 @@ export const plotlyCharts = {
   area,
   bar,
   bubble,
+  csv,
   heatmap,
   hexagons,
   image: slideshow, // both 'image' and 'slideshow' types work for images
@@ -44,12 +45,11 @@ export const plotlyCharts = {
   sankey,
   slideshow,
   text,
+  tile,
   transit,
   scatter,
   vega,
   video,
-  csv,
-  overview,
 }
 
 // export all remaining charts/maps here:
